# openapi.yml
openapi: 3.0.3
info:
  title: "weekend-vocabulary"
  description: |
    This is a server for english vocabulary.
  version: "1.0.0"
paths:
  /get/recent:
    get:
      responses:
        "200":
          description: "Return recently regisitered vocabulary"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecentlyVocabularyResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /get/all:
    get:
      responses:
        "200":
          description: "Return all regisitered vocabulary list"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllVocabularyResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /register:
    put:
      requestBody:
        description: "vocabulary"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterVocabularyRequestBody"
      responses:
        "200":
          description: "Regisiter vocabulary"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterVocabularyOkResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  responses:
    NotFound:
      description: "The specified resource was not found"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: "Bad Request"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalServerError:
      description: "Internal Server Error"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Vocabulary:
      properties:
        id:
          type: integer
        part_of_speech:
          type: string
          enum:
            [
              noun,
              pronoun,
              adjective,
              verb,
              adverb,
              preposition,
              conjunction,
              interjection,
            ]
        spelling:
          type: string
        meaning:
          type: string
      required:
        - part_of_speech
        - spelling
        - meaning
    RecentlyVocabularyResponse:
      properties:
        vocabulary:
          $ref: "#/components/schemas/Vocabulary"
      required:
        - vocabulary
    AllVocabularyResponse:
      properties:
        vocabulary_list:
          type: array
          items:
            $ref: "#/components/schemas/Vocabulary"
        total_count:
          type: integer
      required:
        - vocabulary_list
        - total_count
    RegisterVocabularyInput:
      properties:
        part_of_speech:
          type: string
          enum:
            [
              noun,
              pronoun,
              adjective,
              verb,
              adverb,
              preposition,
              conjunction,
              interjection,
            ]
        spelling:
          type: string
        meaning:
          type: string
      required:
        - part_of_speech
        - spelling
        - meaning
    RegisterVocabularyRequestBody:
      properties:
        vocabulary:
          $ref: "#/components/schemas/RegisterVocabularyInput"
      required:
        - vocabulary
    RegisterVocabularyOkResponse:
      properties:
        message:
          type: string
      required:
        - message
    Error:
      type: object
      properties:
        message:
          type: string